Suppress all RSpec spec file names displayed in stdout at the beginning of running tests in Regular Mode
In Regular Mode, we use RSpec::Core::RakeTask. By default, it prints an RSpec command with all test files and options provided to it. For a big test suite, this can be a large output.
We can suppress that with a verbose false option for the rake task.

Related issue:
https://github.com/KnapsackPro/knapsack_pro-ruby/issues/189