update(queue rspec runner): We ignore configuration.fail_if_no_examples option on purpose
We ignore configuration.fail_if_no_examples option on purpose because it should be off in Queue Mode. RSpec should not fail when examples are not executed for a batch of tests fetched from Queue API. The batch could have tests that have no examples (for example, someone commented out the content of the spec file). We should fetch another batch of tests from Queue API and keep running tests.